Updated to improve response time.
Updated the dixionaryGen function, which should improve response time, and ensured all keys are lower case so the map can just be accessed using a key rather than a linear search and this also saves on having to convert each key to lower case for every word in the message. Ensured the length of the message is greater than 0, which will stop it trying to make a request if the message is an attachment or embed.